rename script files, to avoid namespace collission for cached imported methods
