+++ extensions/pyRevitTools.extension/pyRevit.tab/Modify.panel/edit3.stack/Edit.pulldown/XLS Export.pushbutton/xlsex_config.py
@@ -1,4 +1,4 @@
-from script import main
+from xlsex_script import main
 
 if __name__ == "__main__":
     main(advanced=True)
