+++ extensions/pyRevitTools.extension/pyRevit.tab/Project.panel/ptools.stack/Links.pulldown/Create Workset For Linked Element.pushbutton/ws4links_config.py
@@ -4,7 +4,7 @@ import json
 import os
 from pyrevit import script, forms
 from pyrevit.userconfig import user_config
-from script import main
+from ws4links_script import main
 
 
 def get_translations(script_folder, script_type, locale):
