+++ extensions/pyRevitTools.extension/pyRevit.tab/Project.panel/ptools.stack/Links.pulldown/Create Workset For Linked Element.pushbutton/ws4links_script.py
@@ -87,7 +87,7 @@ def main():
                 title=translations_dict["Worksharing.Enable.Error.Title"],
                 exitscript=True
             )
-        
+
         with revit.Transaction(translations_dict["Transaction.Name"]):
             for el in selection:
                 linked_model_name = ""
